Made necessary changes
