+++ src/populate.js
@@ -1,0 +1,15 @@
+const index = () => {
+  const scores = document.querySelector('.scores');
+
+  scores.innerHTML = `
+  <ul>
+  <li>Name:100</li>
+  <li class="li-grey">Name:10</li>
+  <li>Name:200</li>
+  <li class="li-grey">Name:40</li>
+  <li>Name:60</li>
+  <li class="li-grey">Name:60</li>
+  </ul>`;
+};
+
+export default index;
